+++ kernel/src/boot.rs
@@ -107,7 +107,7 @@ impl fmt::Debug for MemoryRegion {
 #[repr(C, align(8))]
 pub struct Framebuffer {
 	pub base: u64,
-	pub pitch: u16,
+	pub pitch: u32,
 	pub width: u16,
 	pub height: u16,
 	pub bpp: u8,
